Wild Animal Rescue App
This application is made up to help in rescuing wild animals which are lost in Cities. As we have added an Image, GPS, and Message option for the user, the user can take an image, and the GPS will declare his/her location and add a description about the animal in the message section so that the Org saves them could do their task easily.